Factor out cell components
